chore(boards): move utils
